I've created a gui, and moved the main method to a new file called mapSTATS (The name may be changed later when we name the program, but for now this is the name) in a package which contains a graphical interface.
I've also added some images, these are icons for the file menu, which serves basic functions as of late, including selecting a folder using a JFileChooser, and saving the compiled drawn image as an image in the root directory.(bugged, needs fixing). 

Problems: The STD draw method is useless, we need to use the Java.swingx.*; draw methods to do this project, so moving on DO NOT use the simplified draw method for beginners, it is very limited in application. (this is for nigel to read). If you have any issues with this make an issue in the Quantum cloud repo.